+++ python/lsst/ts/m2com/mock/mock_command.py
@@ -20,7 +20,6 @@
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 import asyncio
-import typing
 
 from lsst.ts import salobj
 from lsst.ts.idl.enums import MTM2
@@ -67,7 +66,7 @@ class MockCommand:
 
     async def enable(
         self, message: dict, model: MockModel, message_event: MockMessageEvent
-    ) -> typing.Tuple[MockModel, CommandStatus]:
+    ) -> tuple[MockModel, CommandStatus]:
         """Enable the system.
 
         Parameters
@@ -201,7 +200,7 @@ class MockCommand:
 
     async def disable(
         self, message: dict, model: MockModel, message_event: MockMessageEvent
-    ) -> typing.Tuple[MockModel, CommandStatus]:
+    ) -> tuple[MockModel, CommandStatus]:
         """Disable the system.
 
         Parameters
@@ -240,7 +239,7 @@ class MockCommand:
 
     async def standby(
         self, message: dict, model: MockModel, message_event: MockMessageEvent
-    ) -> typing.Tuple[MockModel, CommandStatus]:
+    ) -> tuple[MockModel, CommandStatus]:
         """Standby the system.
 
         Parameters
@@ -273,7 +272,7 @@ class MockCommand:
 
     async def start(
         self, message: dict, model: MockModel, message_event: MockMessageEvent
-    ) -> typing.Tuple[MockModel, CommandStatus]:
+    ) -> tuple[MockModel, CommandStatus]:
         """Start the system.
 
         Parameters
@@ -305,7 +304,7 @@ class MockCommand:
 
     async def enter_control(
         self, message: dict, model: MockModel, message_event: MockMessageEvent
-    ) -> typing.Tuple[MockModel, CommandStatus]:
+    ) -> tuple[MockModel, CommandStatus]:
         """Enter the control.
 
         This is only supported for the commandable SAL component (CSC). In the
@@ -341,7 +340,7 @@ class MockCommand:
 
     async def exit_control(
         self, message: dict, model: MockModel, message_event: MockMessageEvent
-    ) -> typing.Tuple[MockModel, CommandStatus]:
+    ) -> tuple[MockModel, CommandStatus]:
         """Exit the control.
 
         This is only supported for the commandable SAL component (CSC). In the
@@ -391,7 +390,7 @@ class MockCommand:
 
     async def apply_forces(
         self, message: dict, model: MockModel, message_event: MockMessageEvent
-    ) -> typing.Tuple[MockModel, CommandStatus]:
+    ) -> tuple[MockModel, CommandStatus]:
         """Apply the forces in addtional to the LUT force.
 
         LUT: look-up table.
@@ -425,7 +424,7 @@ class MockCommand:
 
     async def position_mirror(
         self, message: dict, model: MockModel, message_event: MockMessageEvent
-    ) -> typing.Tuple[MockModel, CommandStatus]:
+    ) -> tuple[MockModel, CommandStatus]:
         """Position the mirror.
 
         Parameters
@@ -474,7 +473,7 @@ class MockCommand:
 
     async def reset_force_offsets(
         self, message: dict, model: MockModel, message_event: MockMessageEvent
-    ) -> typing.Tuple[MockModel, CommandStatus]:
+    ) -> tuple[MockModel, CommandStatus]:
         """Reset the actuator force offsets (not LUT force).
 
         LUT: look-up table.
@@ -503,7 +502,7 @@ class MockCommand:
 
     async def clear_errors(
         self, message: dict, model: MockModel, message_event: MockMessageEvent
-    ) -> typing.Tuple[MockModel, CommandStatus]:
+    ) -> tuple[MockModel, CommandStatus]:
         """Clear the system errors.
 
         Parameters
@@ -537,7 +536,7 @@ class MockCommand:
 
     async def switch_force_balance_system(
         self, message: dict, model: MockModel, message_event: MockMessageEvent
-    ) -> typing.Tuple[MockModel, CommandStatus]:
+    ) -> tuple[MockModel, CommandStatus]:
         """Switch the force balance system.
 
         Parameters
@@ -584,7 +583,7 @@ class MockCommand:
 
     async def select_inclination_source(
         self, message: dict, model: MockModel, message_event: MockMessageEvent
-    ) -> typing.Tuple[MockModel, CommandStatus]:
+    ) -> tuple[MockModel, CommandStatus]:
         """Select the source of inclination.
 
         Notes
@@ -621,7 +620,7 @@ class MockCommand:
 
     async def set_temperature_offset(
         self, message: dict, model: MockModel, message_event: MockMessageEvent
-    ) -> typing.Tuple[MockModel, CommandStatus]:
+    ) -> tuple[MockModel, CommandStatus]:
         """Set the temperature offset used in the calculation of LUT force.
 
         LUT: look-up table.
@@ -655,7 +654,7 @@ class MockCommand:
 
     async def switch_command_source(
         self, message: dict, model: MockModel, message_event: MockMessageEvent
-    ) -> typing.Tuple[MockModel, CommandStatus]:
+    ) -> tuple[MockModel, CommandStatus]:
         """Switch the command source to be the commandable SAL component (CSC)
         or engineering user interface (EUI).
 
@@ -683,7 +682,7 @@ class MockCommand:
 
     async def run_script(
         self, message: dict, model: MockModel, message_event: MockMessageEvent
-    ) -> typing.Tuple[MockModel, CommandStatus]:
+    ) -> tuple[MockModel, CommandStatus]:
         """Run the binary script used in the engineering user interface (EUI).
 
         Parameters
@@ -728,7 +727,7 @@ class MockCommand:
 
     async def move_actuators(
         self, message: dict, model: MockModel, message_event: MockMessageEvent
-    ) -> typing.Tuple[MockModel, CommandStatus]:
+    ) -> tuple[MockModel, CommandStatus]:
         """Move the actuators.
 
         Parameters
@@ -773,7 +772,7 @@ class MockCommand:
 
     async def reset_breakers(
         self, message: dict, model: MockModel, message_event: MockMessageEvent
-    ) -> typing.Tuple[MockModel, CommandStatus]:
+    ) -> tuple[MockModel, CommandStatus]:
         """Reset the breakers.
 
         Parameters
@@ -853,7 +852,7 @@ class MockCommand:
 
     async def reboot_controller(
         self, message: dict, model: MockModel, message_event: MockMessageEvent
-    ) -> typing.Tuple[MockModel, CommandStatus]:
+    ) -> tuple[MockModel, CommandStatus]:
         """Reboot the cell controller.
 
         Parameters
@@ -877,7 +876,7 @@ class MockCommand:
 
     async def enable_open_loop_max_limit(
         self, message: dict, model: MockModel, message_event: MockMessageEvent
-    ) -> typing.Tuple[MockModel, CommandStatus]:
+    ) -> tuple[MockModel, CommandStatus]:
         """Enable the maximum limit in open-loop control.
 
         Parameters
@@ -910,7 +909,7 @@ class MockCommand:
 
     async def save_mirror_position(
         self, message: dict, model: MockModel, message_event: MockMessageEvent
-    ) -> typing.Tuple[MockModel, CommandStatus]:
+    ) -> tuple[MockModel, CommandStatus]:
         """Save the position of mirror.
 
         Parameters
@@ -934,7 +933,7 @@ class MockCommand:
 
     async def set_mirror_home(
         self, message: dict, model: MockModel, message_event: MockMessageEvent
-    ) -> typing.Tuple[MockModel, CommandStatus]:
+    ) -> tuple[MockModel, CommandStatus]:
         """Set the home of mirror.
 
         Parameters
@@ -965,7 +964,7 @@ class MockCommand:
 
     async def switch_digital_output(
         self, message: dict, model: MockModel, message_event: MockMessageEvent
-    ) -> typing.Tuple[MockModel, CommandStatus]:
+    ) -> tuple[MockModel, CommandStatus]:
         """Switch the digital output.
 
         Parameters
@@ -1020,7 +1019,7 @@ class MockCommand:
 
     async def power(
         self, message: dict, model: MockModel, message_event: MockMessageEvent
-    ) -> typing.Tuple[MockModel, CommandStatus]:
+    ) -> tuple[MockModel, CommandStatus]:
         """Power on/off the motor/communication system.
 
         Parameters
@@ -1083,7 +1082,7 @@ class MockCommand:
 
     async def reset_actuator_steps(
         self, message: dict, model: MockModel, message_event: MockMessageEvent
-    ) -> typing.Tuple[MockModel, CommandStatus]:
+    ) -> tuple[MockModel, CommandStatus]:
         """Reset the actuator steps.
 
         Notes
@@ -1112,7 +1111,7 @@ class MockCommand:
 
     async def set_closed_loop_control_mode(
         self, message: dict, model: MockModel, message_event: MockMessageEvent
-    ) -> typing.Tuple[MockModel, CommandStatus]:
+    ) -> tuple[MockModel, CommandStatus]:
         """Set the closed-loop control mode.
 
         Parameters
@@ -1144,7 +1143,7 @@ class MockCommand:
 
     async def set_inner_loop_control_mode(
         self, message: dict, model: MockModel, message_event: MockMessageEvent
-    ) -> typing.Tuple[MockModel, CommandStatus]:
+    ) -> tuple[MockModel, CommandStatus]:
         """Set the inner-loop control mode.
 
         Parameters
@@ -1186,7 +1185,7 @@ class MockCommand:
 
     async def get_inner_loop_control_mode(
         self, message: dict, model: MockModel, message_event: MockMessageEvent
-    ) -> typing.Tuple[MockModel, CommandStatus]:
+    ) -> tuple[MockModel, CommandStatus]:
         """get the inner-loop control mode.
 
         Parameters
@@ -1224,7 +1223,7 @@ class MockCommand:
 
     async def load_configuration(
         self, message: dict, model: MockModel, message_event: MockMessageEvent
-    ) -> typing.Tuple[MockModel, CommandStatus]:
+    ) -> tuple[MockModel, CommandStatus]:
         """Load the configuration.
 
         Parameters
@@ -1250,7 +1249,7 @@ class MockCommand:
 
     async def set_control_parameters(
         self, message: dict, model: MockModel, message_event: MockMessageEvent
-    ) -> typing.Tuple[MockModel, CommandStatus]:
+    ) -> tuple[MockModel, CommandStatus]:
         """Set the closed-loop controller (CLC) control parameters.
 
         Parameters