+++ tests/mock/test_mock_model.py
@@ -19,7 +19,6 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-import typing
 import unittest
 
 import numpy as np
@@ -219,7 +218,7 @@ class TestMockModel(unittest.IsolatedAsyncioTestCase):
         tangent_actuator_positions = telemetry_data["tangentEncoderPositions"]
         self.assertEqual(len(tangent_actuator_positions["position"]), NUM_TANGENT_LINK)
 
-    def _apply_forces(self) -> typing.Tuple[list, list]:
+    def _apply_forces(self) -> tuple[list, list]:
         force_axial = [1] * (NUM_ACTUATOR - NUM_TANGENT_LINK)
         force_tangent = [2] * NUM_TANGENT_LINK
         self.model.control_closed_loop.apply_forces(force_axial, force_tangent)
@@ -283,7 +282,7 @@ class TestMockModel(unittest.IsolatedAsyncioTestCase):
 
     def _get_force_error_tangent_expected(
         self,
-    ) -> typing.Tuple[float, numpy.typing.NDArray[np.float64], dict]:
+    ) -> tuple[float, numpy.typing.NDArray[np.float64], dict]:
         return (
             89.853,
             np.array([-325.307, -447.377, 1128.37, -1249.98, 458.63, 267.627]),